+++ packages/core/src/image-manipulation/effects.js
@@ -4,58 +4,6 @@ import * as constants from '../constants';
 import * as compositeModes from './composite-modes';
 
 /**
- * Apply a ordered dithering effect
- * @param {function(Error, Jimp)} cb (optional) a callback for when complete
- * @returns {Jimp} this for chaining of methods
- */
-export function dither565(cb) {
-    const rgb565Matrix = [
-        1,
-        9,
-        3,
-        11,
-        13,
-        5,
-        15,
-        7,
-        4,
-        12,
-        2,
-        10,
-        16,
-        8,
-        14,
-        6
-    ];
-    this.scanQuiet(0, 0, this.bitmap.width, this.bitmap.height, function(
-        x,
-        y,
-        idx
-    ) {
-        const thresholdId = ((y & 3) << 2) + (x % 4);
-        const dither = rgb565Matrix[thresholdId];
-        this.bitmap.data[idx] = Math.min(this.bitmap.data[idx] + dither, 0xff);
-        this.bitmap.data[idx + 1] = Math.min(
-            this.bitmap.data[idx + 1] + dither,
-            0xff
-        );
-        this.bitmap.data[idx + 2] = Math.min(
-            this.bitmap.data[idx + 2] + dither,
-            0xff
-        );
-    });
-
-    if (isNodePattern(cb)) {
-        cb.call(this, null, this);
-    }
-
-    return this;
-}
-
-// alternative reference
-export const dither16 = dither565;
-
-/**
  * Get an image's histogram
  * @return {object} An object with an array of color occurrence counts for each channel (r,g,b)
  */
