+++ packages/jimp/src/index.js
@@ -3,6 +3,7 @@ import configure from '@jimp/custom';
 export default configure({
     types: ['@jimp/jpeg', '@jimp/png', '@jimp/bmp', '@jimp/tiff', '@jimp/gif'],
     plugins: [
+        '@jimp/plugin-dither',
         '@jimp/plugin-resize',
         '@jimp/plugin-blit',
         '@jimp/plugin-rotate',
