+++ packages/plugin-dither/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-dither
+
+dither an image.
